Added test for empty legend.
Added missing location_annotations() to the datasource as the lack of it
caused failing tests. So now the tests function again, but the
legend-is-not-empty test fails still. It needs more mocking (it needs
annotations), but I would really have to dive into the code and @remcogerlich
can probably add it much quicker.